refactor: StrokePaintTool abstract class 로 중복 제거
