+++ app/src/main/java/woowacourse/paint/presentation/ui/canvas/painttool/Eraser.kt
@@ -1,29 +1,18 @@
 package woowacourse.paint.presentation.ui.canvas.painttool
 
-import android.graphics.Paint
 import android.graphics.PorterDuff
 import android.graphics.PorterDuffXfermode
 import woowacourse.paint.presentation.ui.canvas.Palette
 
-class Eraser(
-    palette: Palette,
-) : PaintTool(palette, PorterDuffXfermode(PorterDuff.Mode.CLEAR)) {
+class Eraser(palette: Palette) : StrokePaintTool(palette) {
 
     init {
-        palette.paint.style = Paint.Style.STROKE
+        palette.paint.xfermode = PorterDuffXfermode(PorterDuff.Mode.CLEAR)
     }
 
     override fun nextPath(): PaintTool = Eraser(palette)
 
     override fun changePalette(palette: Palette?): PaintTool {
         return Eraser(palette ?: this.palette.copy())
-    }
-
-    override fun onDownEvent(pointX: Float, pointY: Float) {
-        path.moveTo(pointX, pointY)
-    }
-
-    override fun onMoveEvent(pointX: Float, pointY: Float) {
-        path.lineTo(pointX, pointY)
     }
 }
