+++ app/src/main/java/woowacourse/paint/presentation/ui/canvas/painttool/Line.kt
@@ -1,25 +1,12 @@
 package woowacourse.paint.presentation.ui.canvas.painttool
 
-import android.graphics.Paint
 import woowacourse.paint.presentation.ui.canvas.Palette
 
-class Line(palette: Palette) : PaintTool(palette, null) {
-
-    init {
-        palette.paint.style = Paint.Style.STROKE
-    }
+class Line(palette: Palette) : StrokePaintTool(palette) {
 
     override fun nextPath(): PaintTool = Line(palette)
 
     override fun changePalette(palette: Palette?): PaintTool {
         return Line(palette ?: this.palette.copy())
-    }
-
-    override fun onDownEvent(pointX: Float, pointY: Float) {
-        path.moveTo(pointX, pointY)
-    }
-
-    override fun onMoveEvent(pointX: Float, pointY: Float) {
-        path.lineTo(pointX, pointY)
     }
 }
